A few more fixes needed for GradleRIO
